estable, agregué imagenes para hacer la estimación de cada  historia

+++ db/migrate/20110314180251_add_estimation_to_project.rb
@@ -1,10 +1,9 @@
 class AddEstimationToProject < ActiveRecord::Migration
   def self.up
-    add_column :projects, :estimation, :float    
+    add_column :projects, :estimation, :integer    
   end
 
   def self.down
      remove_column :projects, :public
   end
 end
-
